Use with caution ! This is the modification of Hein's r255, resolving issue 11. All user-configurable settings have been moved into the new config.h file, which replaces ArduUser.h. ArduUser.h is obsolete in this version. A few of the settings from former ArduUser.h have been moved back into the main .pde, since they are not to be user configurable. I did my best to create a walkthrough-like setup file from config.h.
None of the program logic has been altered, it is merely a code clean-up.

I do not intend to have this as a production or test version, mainly a show and tell. If this is the way we think we want to go with this, then that's great of course !